+++ app/src/main/java/com/flyingburritoco/mobilemenu/ViewHolder/CartAdapter.java
@@ -93,7 +93,7 @@ public class CartAdapter extends RecyclerView.Adapter<CartViewHolder>{
 
         Locale locale = new Locale("en","US");
         NumberFormat fmt = NumberFormat.getCurrencyInstance(locale);
-        int price = (Integer.parseInt(listData.get(position).getPrice()))*(Integer.parseInt(listData.get(position).getQuantity()));
+        double price = (Double.parseDouble(listData.get(position).getPrice()))*(Double.parseDouble(listData.get(position).getQuantity()));
         holder.itemPrice.setText(fmt.format(price));
         holder.itemName.setText(listData.get(position).getProductName());
